+++ os/env_test.go
@@ -27,3 +27,20 @@ func TestMustEnv_Panics(t *testing.T) {
 
 	assert.Panics(t, func() { os.MustEnv("test_key") })
 }
+
+func TestDefaultEnv(t *testing.T) {
+	syscall.Clearenv()
+
+	err := syscall.Setenv("key", "value")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	assert.Equal(t, "value", os.DefaultEnv("key", "default"))
+}
+
+func TestDefaultEnv_NotFound(t *testing.T) {
+	syscall.Clearenv()
+
+	assert.Equal(t, "default", os.DefaultEnv("key", "default"))
+}