+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- **os:** `DefaultEnv` to fallback to a default value when a key is not found
+
+
 ## 0.7.0 - 2017-07-27
 
 ### Added